pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "socialcrop"
version = "0.1.0"
description = "Social media image resizer CLI with 15+ platform presets. Zero API, smart crop, batch export."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "fuku" }]
keywords = ["social-media", "image-resize", "cli", "instagram", "twitter", "crop"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",

    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Utilities",
]
dependencies = [
    "Pillow>=10.0",
    "click>=8.1",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-cov>=4.0", "ruff>=0.1"]

[project.scripts]
socialcrop = "socialcrop.cli:main"

[project.urls]
Homepage = "https://github.com/izag8216/socialcrop"
Repository = "https://github.com/izag8216/socialcrop"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
